Migrate spec from w3c/media-source
This update adds the required code of conduct, contributing, license, and w3c
files, and migrates the actual from the w3c/media-source repository.

Changes made to the spec during the migration:
- Switch to latest W3C profile of Respec
- Use the new "group" property to ease switch to Process 2020
- Load JQuery (no longer included in Respec, but needed for media-source.js)
- Update URL of media-source.js script

Note the media-source.js script requires a couple of updates to use an absolute
URL for the ED of the main Media Source (instead of "index.html" which currently
creates links to self instead of creating links to the main MSE spec), as well
as to fix a few broken fragments.